chore(tests): split proxyTable in unit & e2e tests
