+++ test/unit/_libs.js
@@ -1,0 +1,8 @@
+module.exports = {
+    configFactory: require('../../lib/config-factory'),
+    contextMatcher: require('../../lib/context-matcher'),
+    handlers: require('../../lib/handlers'),
+    Logger: require('../../lib/logger'),
+    pathRewriter: require('../../lib/path-rewriter'),
+    proxyTable: require('../../lib/proxy-table')
+};
